refactor: extract tar utilities, add content size limit and a11y fixes
- Extract createSingleFileTar and extractFileFromTar into shared
  apps/web/src/lib/tar.ts with consistent ustar format
- Remove duplicated tar implementations from instructions route,
  docker-sandbox, and test helper
- Add 1MB content size limit to PUT instructions endpoint (413 response)
- Add aria-label to instructions textarea for screen readers
- Add comment explaining S3-before-DB upload design decision
