+++ turbo/apps/web/app/api/agent/composes/[id]/instructions/__tests__/route.test.ts
@@ -16,52 +16,10 @@ import {
   MOCK_USER_EMAIL,
 } from "../../../../../../../src/__tests__/clerk-mock";
 import { getInstructionsStorageName } from "@vm0/core";
+import { createSingleFileTar } from "../../../../../../../src/lib/tar";
 
-/**
- * Build a gzipped tar archive containing a single file.
- * The route downloads archive.tar.gz, decompresses, and extracts from tar.
- */
 function buildTarGz(filename: string, content: string): Buffer {
-  const contentBuf = Buffer.from(content, "utf-8");
-
-  // Tar header: 512 bytes
-  const header = Buffer.alloc(512);
-  // File name at offset 0 (up to 100 bytes)
-  header.write(filename, 0, Math.min(filename.length, 100), "utf-8");
-  // File mode at offset 100 (8 bytes, octal)
-  header.write("0000644\0", 100, 8, "utf-8");
-  // Owner/group IDs at offsets 108, 116 (8 bytes each, octal)
-  header.write("0000000\0", 108, 8, "utf-8");
-  header.write("0000000\0", 116, 8, "utf-8");
-  // File size at offset 124 (12 bytes, octal, null-terminated)
-  const sizeOctal = contentBuf.length.toString(8).padStart(11, "0");
-  header.write(sizeOctal + "\0", 124, 12, "utf-8");
-  // Modification time at offset 136 (12 bytes, octal)
-  const mtime = Math.floor(Date.now() / 1000)
-    .toString(8)
-    .padStart(11, "0");
-  header.write(mtime + "\0", 136, 12, "utf-8");
-  // Type flag at offset 156: '0' = regular file
-  header.write("0", 156, 1, "utf-8");
-
-  // Compute checksum: fill checksum field (offset 148, 8 bytes) with spaces first
-  header.write("        ", 148, 8, "utf-8");
-  let checksum = 0;
-  for (let i = 0; i < 512; i++) {
-    checksum += header[i] ?? 0;
-  }
-  // Write checksum as 6-digit octal, null-terminated, space-padded
-  header.write(checksum.toString(8).padStart(6, "0") + "\0 ", 148, 8, "utf-8");
-
-  // Pad content to 512-byte boundary
-  const paddingSize = (512 - (contentBuf.length % 512)) % 512;
-  const padding = Buffer.alloc(paddingSize);
-
-  // End-of-archive marker: two 512-byte zero blocks
-  const endMarker = Buffer.alloc(1024);
-
-  const tar = Buffer.concat([header, contentBuf, padding, endMarker]);
-  return gzipSync(tar);
+  return gzipSync(createSingleFileTar(filename, Buffer.from(content, "utf-8")));
 }
 
 const context = testContext();
